backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
pyjwt
requests
aiofiles
